refactor agent-core loop and code mode host modules
